+++ pkg/querier/batch/merge.go
@@ -3,11 +3,26 @@ package batch
 import (
 	"container/heap"
 	"sort"
+	"sync"
 
 	"github.com/prometheus/prometheus/tsdb/chunkenc"
 
 	promchunk "github.com/cortexproject/cortex/pkg/chunk"
 )
+
+// poolBatchesBuf controls whether buildNextBatch uses sync.Pool or per-iterator batchesBuf.
+var poolBatchesBuf bool
+
+func SetPoolBatchesBuf(enabled bool) {
+	poolBatchesBuf = enabled
+}
+
+var batchesBufPool = sync.Pool{
+	New: func() any {
+		buf := make(batchStream, 3)
+		return &buf
+	},
+}
 
 type mergeIterator struct {
 	its []*nonOverlappingIterator
@@ -19,6 +34,9 @@ type mergeIterator struct {
 	// Buffers to merge in.
 	batchesBuf   batchStream
 	nextBatchBuf [1]promchunk.Batch
+
+	numPartitions int
+	usePool       bool
 
 	currErr error
 }
@@ -32,9 +50,14 @@ func newMergeIterator(it iterator, cs []GenericChunk) *mergeIterator {
 		c = mIterator.Reset(len(css))
 	} else {
 		c = &mergeIterator{
-			h:          make(iteratorHeap, 0, len(css)),
-			batches:    make(batchStream, 0, len(css)),
-			batchesBuf: make(batchStream, len(css)),
+			h:             make(iteratorHeap, 0, len(css)),
+			batches:       make(batchStream, 0, len(css)),
+			numPartitions: len(css),
+			usePool:       poolBatchesBuf,
+		}
+
+		if !poolBatchesBuf {
+			c.batchesBuf = make(batchStream, len(css))
 		}
 	}
 
@@ -65,13 +88,16 @@ func (c *mergeIterator) Reset(size int) *mergeIterator {
 	c.its = c.its[:0]
 	c.h = c.h[:0]
 	c.batches = c.batches[:0]
+	c.numPartitions = size
 
-	if size > cap(c.batchesBuf) {
-		c.batchesBuf = make(batchStream, len(c.its))
-	} else {
-		c.batchesBuf = c.batchesBuf[:size]
-		for i := range size {
-			c.batchesBuf[i] = promchunk.Batch{}
+	if !c.usePool {
+		if size > cap(c.batchesBuf) {
+			c.batchesBuf = make(batchStream, size)
+		} else {
+			c.batchesBuf = c.batchesBuf[:size]
+			for i := range size {
+				c.batchesBuf[i] = promchunk.Batch{}
+			}
 		}
 	}
 
@@ -141,18 +167,48 @@ func (c *mergeIterator) nextBatchEndTime() int64 {
 }
 
 func (c *mergeIterator) buildNextBatch(size int) chunkenc.ValueType {
+	if len(c.h) == 0 && len(c.batches) > 0 {
+		return c.batches[0].ValType
+	}
+	if len(c.h) == 0 {
+		return chunkenc.ValNone
+	}
+
+	var buf batchStream
+	if c.usePool {
+		bp := batchesBufPool.Get().(*batchStream)
+		defer func() {
+			for i := range *bp {
+				(*bp)[i] = promchunk.Batch{}
+			}
+			batchesBufPool.Put(bp)
+		}()
+		if cap(*bp) < c.numPartitions {
+			*bp = make(batchStream, c.numPartitions)
+		} else {
+			*bp = (*bp)[:c.numPartitions]
+		}
+		buf = *bp
+	} else {
+		buf = c.batchesBuf
+	}
+
 	// All we need to do is get enough batches that our first batch's last entry
 	// is before all iterators next entry.
 	for len(c.h) > 0 && (len(c.batches) == 0 || c.nextBatchEndTime() >= c.h[0].AtTime()) {
 		c.nextBatchBuf[0] = c.h[0].Batch()
-		c.batchesBuf = mergeStreams(c.batches, c.nextBatchBuf[:], c.batchesBuf, size)
-		c.batches = append(c.batches[:0], c.batchesBuf...)
+		buf = mergeStreams(c.batches, c.nextBatchBuf[:], buf, size)
+		c.batches = append(c.batches[:0], buf...)
 
 		if valType := c.h[0].Next(size); valType != chunkenc.ValNone {
 			heap.Fix(&c.h, 0)
 		} else {
 			heap.Pop(&c.h)
 		}
+	}
+
+	if !c.usePool {
+		c.batchesBuf = buf
 	}
 
 	if len(c.batches) > 0 {